refactor(agent_supervisor/lifecycle): split tests to tests_lifecycle.rs (#1266)
- src/agent_supervisor/lifecycle.rs: 794 → 469 LOC
- New src/agent_supervisor/tests_lifecycle.rs (325 LOC)
- Includes both inline 'mod tests' body AND the two zombie-reaper test mods

mod.rs adjusted to register tests_lifecycle (existing tests.rs / tests_tmux.rs unchanged).
Pure refactor: 14 tests pass.
